Everything is now none static.
All Hard coded puzzles are now located in HCPuzzles.java

Deleted backtracking.java as it is implemented in Solver.java now

Generator.java has a recursive and a none recursive method to generate puzzles with a specified amount of numbers taken out.

A method to find every solution for a given puzzle has been made.